api/requirements.txt
fastapi
uvicorn[standard]
pandas
python-dotenv
fastparquet